Rename, reorganize and refactor editors & related
