+++ react/components/EditorContainer/Sidebar/ConfigurationList/Editor/ConditionsSelector/ScopeSelector.tsx
