+++ react/components/EditorContainer/Sidebar/ConfigurationList/Editor/ConditionsSelector/index.tsx
