+++ react/components/EditorContainer/Sidebar/ConfigurationList/List/Card.tsx
